Move shared components to components/ui/
