+++ client/src/components/ui/button-link.tsx
@@ -5,11 +5,12 @@ import { twMerge } from "tailwind-merge"
 
 type Props = React.ComponentPropsWithRef<"a">
 
+// TODO: theme
+
 function Link({ className, ...props }: Readonly<Props>) {
     return (
         <a
             className={twMerge(
-                // TODO: theme
                 `block w-full cursor-pointer truncate rounded-full px-8 py-2 text-center font-medium transition
                 select-none hover:opacity-90 disabled:cursor-not-allowed disabled:opacity-75`,
                 className
