+++ client/src/components/ui/card.tsx
@@ -4,8 +4,9 @@ import { twMerge } from "tailwind-merge"
 
 type Props = React.ComponentPropsWithoutRef<"div">
 
+// TODO: theme
+
 export function Card({ className, ...props }: Readonly<Props>) {
-    // TODO: theme
     return (
         <div
             className={twMerge("h-fit w-full rounded-lg bg-white p-4 shadow-md", className)}
