+++ client/src/components/ui/echo-logo.tsx
