+++ client/src/features/auth/layout/layout.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 
-import { MotionContainer } from "../../../components/container"
-import { Page } from "../../../components/page"
+import { MotionContainer } from "../../../components/ui/container"
+import { Page } from "../../../components/ui/page"
 import { BackToHome } from "../components/back-to-home"
 
 interface Props {
