+++ client/src/features/landing/layout/layout.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 
-import { Page } from "../../../components/page"
+import { Page } from "../../../components/ui/page"
 
 interface Props {
     children: React.ReactNode
@@ -8,7 +8,7 @@ interface Props {
 
 export function Layout({ children }: Readonly<Props>) {
     return (
-        <Page className="flex items-center justify-center bg-gradient-to-br">
+        <Page className="to-echo-teal from-echo-navy flex items-center justify-center bg-gradient-to-br">
             <section className="w-full max-w-md">{children}</section>
         </Page>
     )
